[refactor] 상품 삭제 alert 분리
